# internal/web/server.go
package web

import (
	"context"
	"crypto/rand"
	"crypto/sha256"
	"encoding/hex"
	"fmt"
	"log/slog"
	"net/http"
	"regexp"
	"runtime/debug"
	"strconv"
	"strings"
	"time"

	"github.com/dodopok/estevao-api-go/internal/db"
	"github.com/dodopok/estevao-api-go/internal/rb"
)

// HandlerFunc is a controller action.
type HandlerFunc func(c *Context)

// Endpoint describes how an action is dispatched.
type Endpoint struct {
	Handler HandlerFunc
	// Application marks ApplicationController descendants: they get the
	// JSON headers filter and the production rescue_from StandardError.
	Application bool
	// Plain skips the controller default headers (ActionController::API
	// subclasses not inheriting ApplicationController still get them; only
	// Rack-level endpoints set this).
	Plain bool
	// Live marks ActionController::Live controllers: their streamed
	// responses carry none of the controller default headers, skip
	// Rack::ETag, Runtime and RequestId, and are sent chunked.
	Live bool
}

// Response is a finished HTTP response inside the middleware chain.
type Response struct {
	Status int
	Header http.Header
	Body   []byte
	// Static responses are served before Runtime/RequestId (no ids added).
	Static bool
	// Live responses are streamed (see Endpoint.Live).
	Live    bool
	KeepIDs bool
	// NoETag skips Rack::ETag (a Rack app answering with a String status).
	NoETag bool
	// KeepTypeOn304 keeps Content-Type on a 304: Rack::ConditionalGet
	// deletes "content-type", which misses a Rack app's mixed-case header.
	KeepTypeOn304 bool
	// HeadLength answers HEAD with the body's length (Rack::Files).
	HeadLength bool
}

// Middleware may answer before the application (Rack::Attack).
type Middleware func(r *http.Request) *Response

// Server is the Rails-compatible HTTP stack.
type Server struct {
	Router    *Router
	Endpoints map[string]Endpoint
	Attack    Middleware
	Static    func(r *http.Request) *Response
	// Mount answers for Rack apps mounted ahead of the routes (nil cascades).
	Mount  func(r *http.Request) *Response
	CORS   *CORS
	Logger *slog.Logger
	// ReportError receives unexpected failures (the New Relic hook).
	ReportError func(c *Context, err any, stack []byte)
}

var defaultHeaders = [][2]string{
	{"X-Frame-Options", "SAMEORIGIN"},
	{"X-Xss-Protection", "0"},
	{"X-Content-Type-Options", "nosniff"},
	{"X-Permitted-Cross-Domain-Policies", "none"},
	{"Referrer-Policy", "strict-origin-when-cross-origin"},
}

var requestIDSanitizer = regexp.MustCompile(`[^\w\-@]`)

func newUUID() string {
	var b [16]byte
	_, _ = rand.Read(b[:])
	b[6] = (b[6] & 0x0f) | 0x40
	b[8] = (b[8] & 0x3f) | 0x80
	h := hex.EncodeToString(b[:])
	return h[0:8] + "-" + h[8:12] + "-" + h[12:16] + "-" + h[16:20] + "-" + h[20:]
}

// ServeHTTP runs the stack: Cors > SSL > Runtime > RequestId >
// ShowExceptions > Head > ConditionalGet > ETag > Attack > routes.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	start := time.Now()
	if s.CORS != nil {
		if resp := s.CORS.Preflight(r); resp != nil {
			writeResponse(w, r, resp)
			return
		}
	}
	if s.Static != nil {
		if resp := s.Static(r); resp != nil {
			resp.Header.Set("Strict-Transport-Security", "max-age=63072000; includeSubDomains")
			if s.CORS != nil {
				s.CORS.Decorate(r, resp)
			}
			writeResponse(w, r, resp)
			return
		}
	}
	reqID := newUUID()
	if raw := r.Header.Get("X-Request-Id"); strings.TrimSpace(raw) != "" {
		reqID = requestIDSanitizer.ReplaceAllString(raw, "")
		if len(reqID) > 255 {
			reqID = reqID[:255]
		}
	}
	r = r.WithContext(context.WithValue(r.Context(), routeFormatKey{}, new(string)))
	resp := s.showExceptions(r, reqID)
	if !resp.Live || resp.KeepIDs || resp.Status == 304 {
		resp.Header.Set("X-Request-Id", reqID)
		resp.Header.Set("X-Runtime", fmt.Sprintf("%0.6f", time.Since(start).Seconds()))
	}
	resp.Header.Set("Strict-Transport-Security", "max-age=63072000; includeSubDomains")
	if s.CORS != nil {
		s.CORS.Decorate(r, resp)
	}
	writeResponse(w, r, resp)
}

func writeResponse(w http.ResponseWriter, r *http.Request, resp *Response) {
	h := w.Header()
	for k, v := range resp.Header {
		h[k] = v
	}
	// Puma sends no Date header; Rails adds one only through expires_in.
	// A nil entry stops net/http from adding its own.
	if _, ok := h["Date"]; !ok {
		h["Date"] = nil
	}
	if r.Method == http.MethodHead {
		if resp.HeadLength {
			h.Set("Content-Length", strconv.Itoa(len(resp.Body)))
		} else {
			h.Set("Content-Length", "0")
		}
		w.WriteHeader(resp.Status)
		return
	}
	if resp.Status == 204 || resp.Status == 304 || (resp.Status >= 100 && resp.Status < 200) {
		if h.Get("Content-Length") != "0" {
			h.Del("Content-Length")
		}
		w.WriteHeader(resp.Status)
		return
	}
	if resp.Live {
		h.Del("Content-Length")
		w.WriteHeader(resp.Status)
		_, _ = w.Write(resp.Body)
		if f, ok := w.(http.Flusher); ok {
			f.Flush()
		}
		return
	}
	h.Set("Content-Length", strconv.Itoa(len(resp.Body)))
	w.WriteHeader(resp.Status)
	_, _ = w.Write(resp.Body)
}

// rescueResponses is the part of ActionDispatch::ExceptionWrapper's
// rescue_responses the ported controllers can raise.
var rescueResponses = map[string]int{
	"ActiveRecord::RecordNotFound":                 404,
	"ActionController::ParameterMissing":           400,
	"ActionController::BadRequest":                 400,
	"ActionDispatch::Http::Parameters::ParseError": 400,
	"ActiveRecord::RecordInvalid":                  422,
	"ActiveRecord::RecordNotSaved":                 422,
	"ActionController::InvalidAuthenticityToken":   422,
	"ActionController::UnknownHttpMethod":          405,
	"ActionController::MethodNotAllowed":           405,
	"ActionController::NotImplemented":             501,
}

// showExceptions renders errors escaping the inner stack through the
// public exceptions app.
func (s *Server) showExceptions(r *http.Request, reqID string) (resp *Response) {
	defer func() {
		if rec := recover(); rec != nil {
			status := 500
			if es, ok := rec.(exceptionStatus); ok {
				status = int(es)
			} else if se, ok := rec.(*StandardError); ok && rescueResponses[se.Class] != 0 {
				status = rescueResponses[se.Class]
			} else if s.Logger != nil {
				s.Logger.Error("unhandled error", "error", fmt.Sprint(rec), "request_id", reqID, "stack", string(debug.Stack()))
			}
			resp = PublicException(r, status)
		}
	}()
	resp = s.conditional(r, reqID)
	return resp
}

// conditional applies Rack::ETag then Rack::ConditionalGet.
func (s *Server) conditional(r *http.Request, reqID string) *Response {
	resp := s.inner(r, reqID)
	var digest string
	if !resp.Live && !resp.NoETag && (resp.Status == 200 || resp.Status == 201) && resp.Header.Get("Etag") == "" && resp.Header.Get("Last-Modified") == "" && len(resp.Body) > 0 {
		sum := sha256.Sum256(resp.Body)
		digest = hex.EncodeToString(sum[:])[:32]
		resp.Header.Set("Etag", `W/"`+digest+`"`)
	}
	if resp.Header.Get("Cache-Control") == "" {
		if digest != "" {
			resp.Header.Set("Cache-Control", "max-age=0, private, must-revalidate")
		} else {
			resp.Header.Set("Cache-Control", "no-cache")
		}
	}
	if (r.Method == http.MethodGet || r.Method == http.MethodHead) && resp.Status == 200 {
		if fresh(r, resp.Header) {
			resp.Status = 304
			resp.Live = false
			if !resp.KeepTypeOn304 {
				resp.Header.Del("Content-Type")
			}
			resp.Header.Del("Content-Length")
			resp.Body = nil
		}
	}
	return resp
}

func fresh(r *http.Request, h http.Header) bool {
	if nm, ok := r.Header["If-None-Match"]; ok {
		return h.Get("Etag") == strings.Join(nm, ", ")
	}
	if ims := r.Header.Get("If-Modified-Since"); ims != "" {
		since, err := http.ParseTime(ims)
		if err != nil {
			return false
		}
		lm := h.Get("Last-Modified")
		if lm == "" {
			return false
		}
		last, err := http.ParseTime(lm)
		if err != nil {
			return false
		}
		return !since.Before(last)
	}
	return false
}

func (s *Server) inner(r *http.Request, reqID string) *Response {
	if s.Attack != nil {
		if resp := s.Attack(r); resp != nil {
			return resp
		}
	}
	if s.Mount != nil {
		if resp := s.Mount(r); resp != nil {
			return resp
		}
	}
	m, ok := s.Router.Recognize(r.Method, r.URL.EscapedPath())
	if !ok {
		panic(exceptionStatus(404))
	}
	if holder, ok := r.Context().Value(routeFormatKey{}).(*string); ok {
		*holder = m.Format
	}
	ep, ok := s.Endpoints[m.Endpoint]
	if !ok {
		// A route exists in Rails whose controller is not part of the
		// ported surface (framework engines). Answer as Rails does for an
		// unreachable controller.
		panic(exceptionStatus(404))
	}
	return s.dispatch(r, reqID, m, ep)
}

func (s *Server) dispatch(r *http.Request, reqID string, m *Match, ep Endpoint) *Response {
	controller, action, _ := strings.Cut(m.Endpoint, "#")
	c := &Context{
		R:          r,
		Ctx:        r.Context(),
		RequestID:  reqID,
		Endpoint:   m.Endpoint,
		PathParams: m.Params,
		Format:     m.Format,
		Controller: controller,
		Action:     action,
		Status:     200,
		Header:     http.Header{},
		Values:     map[string]any{},
	}
	if !ep.Plain && !ep.Live {
		for _, kv := range defaultHeaders {
			c.Header.Set(kv[0], kv[1])
		}
	}
	if ep.Application {
		c.Header.Set("Content-Type", "application/json; charset=utf-8")
		c.Header.Set("Content-Disposition", "inline")
		c.Header.Set("X-Request-Id", reqID)
	}
	s.run(c, ep)
	// ActionDispatch::Response#handle_conditional_get!: a response that
	// already carries a validator defaults to must-revalidate (Rack::ETag
	// would otherwise say no-cache).
	if !ep.Live && c.Header.Get("Cache-Control") == "" && (c.Header.Get("Etag") != "" || c.Header.Get("Last-Modified") != "") {
		c.Header.Set("Cache-Control", "max-age=0, private, must-revalidate")
	}
	if !c.written {
		c.Status = 204
		c.Body = nil
	}
	if c.written && c.Body != nil && c.varyAccept() && c.Header.Get("Vary") == "" {
		c.Header.Set("Vary", "Accept")
	}
	if c.Status == 204 {
		// No-content responses go out without a Content-Type (the one
		// set_json_headers put there does not survive).
		c.Header.Del("Content-Type")
	}
	return &Response{Status: c.Status, Header: c.Header, Body: c.Body, Live: ep.Live, KeepIDs: c.KeepRequestIDs}
}

func (s *Server) run(c *Context, ep Endpoint) {
	defer func() {
		rec := recover()
		if rec == nil {
			return
		}
		if _, halted := rec.(haltSignal); halted {
			s.afterAction(c, ep)
			return
		}
		// An exception leaving the action meets the same ensure: the ParseError
		// it raises replaces the exception.
		if _, isStatus := rec.(exceptionStatus); !isStatus && ep.Application && c.params == nil && jsonContentTypes[c.contentType()] {
			if pe := c.parseError(); pe != nil {
				rec = pe
			}
		}
		switch e := rec.(type) {
		case exceptionStatus:
			panic(e)
		case *DomainError:
			if !ep.Application {
				panic(e)
			}
			if c.RenderError != nil {
				c.RenderError(c, e.Class, e.Code, e.Message, e.Context, e.Status())
				return
			}
			c.JSON(e.Status(), rb.M("error", e.Message, "code", e.Code, "request_id", c.RequestID))
			return
		case *InfraError:
			if !ep.Application {
				panic(e)
			}
			if !definesCurrentUser(c.Controller) {
				panic(exceptionStatus(500))
			}
			if s.ReportError != nil {
				s.ReportError(c, e, nil)
			}
			if c.RenderError != nil {
				c.RenderError(c, e.Class, e.Code, e.Message, nil, e.Status())
				return
			}
			c.JSON(e.Status(), rb.M("error", e.Message, "code", e.Code, "request_id", c.RequestID))
			return
		}
		if !ep.Application {
			panic(rec)
		}
		stack := debug.Stack()
		if s.ReportError != nil {
			s.ReportError(c, rec, stack)
		}
		if !definesCurrentUser(c.Controller) {
			if s.Logger != nil {
				s.Logger.Error("[ERROR]", "error", fmt.Sprint(rec), "endpoint", c.Endpoint, "request_id", c.RequestID, "stack", string(stack))
			}
			panic(exceptionStatus(500))
		}
		if s.Logger != nil {
			s.Logger.Error("[ERROR]", "error", fmt.Sprint(rec), "endpoint", c.Endpoint, "request_id", c.RequestID, "stack", string(stack))
		}
		msg := fmt.Sprint(rec)
		if se, ok := rec.(*StandardError); ok {
			msg = se.Message
		} else if err, ok := rec.(error); ok {
			msg = err.Error()
			// An unexpected PostgreSQL error reads as ActiveRecord reports it.
			if _, pgMsg, isPG := db.RubyError(err); isPG {
				msg = pgMsg
			}
		}
		c.JSON(500, rb.M("error", "Internal server error", "message", msg, "trace_id", nil))
	}()
	ep.Handler(c)
	s.afterAction(c, ep)
}

// parseError is the ParseError reading params would raise, if any.
func (c *Context) parseError() (pe *StandardError) {
	defer func() {
		if rec := recover(); rec != nil {
			if se, ok := rec.(*StandardError); ok && se.Class == "ActionDispatch::Http::Parameters::ParseError" {
				pe = se
				return
			}
			panic(rec)
		}
	}()
	c.BodyParams()
	return nil
}

// afterAction ports the ensure of ApplicationController#track_request_timing,
// which reads params[:office_type] after every action. For a JSON body that
// does not parse, that read raises ParseError once the action has already
// rendered; rescue_from then renders again (DoubleRenderError), which only
// the public exceptions app answers: a bare 500.
func (s *Server) afterAction(c *Context, ep Endpoint) {
	if !ep.Application || !jsonContentTypes[c.contentType()] || c.params != nil {
		return
	}
	func() {
		defer func() {
			if rec := recover(); rec != nil {
				if se, ok := rec.(*StandardError); ok && se.Class == "ActionDispatch::Http::Parameters::ParseError" {
					panic(exceptionStatus(500))
				}
				panic(rec)
			}
		}()
		c.Params()
	}()
}

// varyAccept ports ActionDispatch::Request#should_apply_vary_header?.
func (c *Context) varyAccept() bool {
	if c.Format != "" || rb.Present(c.QueryParams().Get("format")) {
		return false
	}
	accept := c.R.Header.Get("Accept")
	xhr := strings.EqualFold(c.R.Header.Get("X-Requested-With"), "XMLHttpRequest")
	if xhr && (strings.TrimSpace(accept) != "" || c.R.Header.Get("Content-Type") != "") {
		return true
	}
	return strings.TrimSpace(accept) != "" && !browserLikeAccepts.MatchString(accept)
}

var browserLikeAccepts = regexp.MustCompile(`,\s*\*/\*|\*/\*\s*,`)

// noCurrentUser are the controllers that include neither Authenticatable
// nor a current_user of their own. NewRelicErrorTracking's handlers
// (handle_standard_error, and report_error_to_newrelic in
// handle_infrastructure_error) read current_user, so in these controllers
// the handler itself raises NameError and the exception reaches the public
// exceptions app: a bare 500.
var noCurrentUser = map[string]bool{
	"api/v1/developers": true, "api/v1/developers/api_keys": true, "api/v1/developers/billing": true,
	"api/v1/plans": true, "api/v1/preferences": true, "api/v1/webhooks/stripe": true, "home": true, "health": true,
	"api/v2/days": true, "api/v2/readings": true, "api/v2/liturgical_explanations": true, "api/v2/collects": true,
	"api/v2/passages": true, "api/v2/prayer_books": true, "api/v2/bible_versions": true, "api/v2/celebrations": true,
	"api/v2/years": true,
}

func definesCurrentUser(controller string) bool { return !noCurrentUser[controller] }